+++ src/infra/files/filesCmdRepo_test.go
@@ -30,7 +30,9 @@ func TestFilesCmdRepo(t *testing.T) {
 	})
 
 	t.Run("CreateUnixFile", func(t *testing.T) {
-		filePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir/filesCmdRepoTest.txt")
+		filePath, _ := valueObject.NewUnixFilePath(
+			fileBasePathStr + "/testDir/filesCmdRepoTest.txt",
+		)
 		mimeType, _ := valueObject.NewMimeType("generic")
 
 		dto := dto.NewCreateUnixFile(filePath, &filePermissions, mimeType)
@@ -42,10 +44,14 @@ func TestFilesCmdRepo(t *testing.T) {
 	})
 
 	t.Run("UpdateUnixFileContent", func(t *testing.T) {
-		filePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir/filesCmdRepoTest.txt")
+		filePath, _ := valueObject.NewUnixFilePath(
+			fileBasePathStr + "/testDir/filesCmdRepoTest.txt",
+		)
 		encodedContent, _ := valueObject.NewEncodedContent("Q29udGVudCB0byB0ZXN0")
 
-		err := filesCmdRepo.UpdateContent(filePath, encodedContent)
+		updateContentDto := dto.NewUpdateUnixFileContent(filePath, encodedContent)
+
+		err := filesCmdRepo.UpdateContent(updateContentDto)
 		if err != nil {
 			t.Errorf("UnexpectedError: %v", err)
 		}
@@ -55,7 +61,11 @@ func TestFilesCmdRepo(t *testing.T) {
 		filePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir")
 		filePermissions, _ := valueObject.NewUnixFilePermissions("0777")
 
-		err := filesCmdRepo.UpdatePermissions(filePath, filePermissions)
+		updatePermissionsDto := dto.NewUpdateUnixFilePermissions(
+			filePath, filePermissions,
+		)
+
+		err := filesCmdRepo.UpdatePermissions(updatePermissionsDto)
 		if err != nil {
 			t.Errorf("UnexpectedError: %v", err)
 		}
@@ -66,7 +76,11 @@ func TestFilesCmdRepo(t *testing.T) {
 			fileBasePathStr + "/testDir/filesCmdRepoTest.txt",
 		)
 
-		err := filesCmdRepo.UpdatePermissions(filePath, filePermissions)
+		updatePermissionsDto := dto.NewUpdateUnixFilePermissions(
+			filePath, filePermissions,
+		)
+
+		err := filesCmdRepo.UpdatePermissions(updatePermissionsDto)
 		if err != nil {
 			t.Errorf("UnexpectedError: %v", err)
 		}
@@ -74,9 +88,13 @@ func TestFilesCmdRepo(t *testing.T) {
 
 	t.Run("MoveUnixDirectory", func(t *testing.T) {
 		sourceFilePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir")
-		destinationFilePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir_")
+		destinationFilePath, _ := valueObject.NewUnixFilePath(
+			fileBasePathStr + "/testDir_",
+		)
 
-		err := filesCmdRepo.Move(sourceFilePath, destinationFilePath, true)
+		moveDto := dto.NewMoveUnixFile(sourceFilePath, destinationFilePath, true)
+
+		err := filesCmdRepo.Move(moveDto)
 		if err != nil {
 			t.Errorf("UnexpectedError: %v", err)
 		}
@@ -90,7 +108,9 @@ func TestFilesCmdRepo(t *testing.T) {
 			fileBasePathStr + "/filesCmdRepoTest.txt",
 		)
 
-		err := filesCmdRepo.Move(sourceFilePath, destinationFilePath, false)
+		moveDto := dto.NewMoveUnixFile(sourceFilePath, destinationFilePath, false)
+
+		err := filesCmdRepo.Move(moveDto)
 		if err != nil {
 			t.Errorf("UnexpectedError: %v", err)
 		}
@@ -98,7 +118,9 @@ func TestFilesCmdRepo(t *testing.T) {
 
 	t.Run("CopyUnixDirectory", func(t *testing.T) {
 		sourceFilePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir_")
-		destinationFilePath, _ := valueObject.NewUnixFilePath(fileBasePathStr + "/testDir")
+		destinationFilePath, _ := valueObject.NewUnixFilePath(
+			fileBasePathStr + "/testDir",
+		)
 
 		dto := dto.NewCopyUnixFile(sourceFilePath, destinationFilePath, true)