+++ src/infra/files/filesQueryRepo_test.go
@@ -13,23 +13,23 @@ func TestFilesQueryRepo(t *testing.T) {
 	currentUser, _ := user.Current()
 	userHomeDir := "/home/" + currentUser.Username
 
-	t.Run("GetFiles", func(t *testing.T) {
+	t.Run("Read", func(t *testing.T) {
 		unixDirPath, _ := valueObject.NewUnixFilePath(userHomeDir)
-		_, err := filesQueryRepo.Get(unixDirPath)
+		_, err := filesQueryRepo.Read(unixDirPath)
 		if err != nil {
 			t.Errorf("ExpectedNoErrorButGot: %s", err.Error())
 		}
 	})
 
-	t.Run("GetFilesWithInvalidDirectory", func(t *testing.T) {
+	t.Run("ReadWithInvalidDirectory", func(t *testing.T) {
 		invalidUnixPath, _ := valueObject.NewUnixFilePath("/aaa/bbb/ccc")
-		_, err := filesQueryRepo.Get(invalidUnixPath)
+		_, err := filesQueryRepo.Read(invalidUnixPath)
 		if err == nil {
 			t.Errorf("ExpectedErrorButGotNil")
 		}
 	})
 
-	t.Run("GetFilesFollowingSymlink", func(t *testing.T) {
+	t.Run("ReadFollowingSymlink", func(t *testing.T) {
 		downloadsDirPath, _ := valueObject.NewUnixFilePath(userHomeDir + "/Downloads")
 		tmpSymlinkPath, _ := valueObject.NewUnixFilePath(userHomeDir + "/tmpSymlink")
 
@@ -40,7 +40,7 @@ func TestFilesQueryRepo(t *testing.T) {
 			t.Errorf("ExpectedNoErrorButGot: %s", err.Error())
 		}
 
-		files, err := filesQueryRepo.Get(tmpSymlinkPath)
+		files, err := filesQueryRepo.Read(tmpSymlinkPath)
 		if err != nil {
 			t.Errorf("ExpectedNoErrorButGot: %s", err.Error())
 		}
@@ -51,9 +51,9 @@ func TestFilesQueryRepo(t *testing.T) {
 		_ = infraHelper.RemoveSymlink(tmpSymlinkPath.String())
 	})
 
-	t.Run("GetSingleFile", func(t *testing.T) {
+	t.Run("ReadFirstFile", func(t *testing.T) {
 		unixFilePath, _ := valueObject.NewUnixFilePath(userHomeDir + "/.bashrc")
-		_, err := filesQueryRepo.GetOne(unixFilePath)
+		_, err := filesQueryRepo.ReadFirst(unixFilePath)
 		if err != nil {
 			t.Errorf("ExpectedNoErrorButGot: %s", err.Error())
 		}